Update gateway index view to use a table with groups
