+++ elixir/apps/web/lib/web/components/core_components.ex
@@ -94,7 +94,10 @@ defmodule Web.CoreComponents do
         <%= for tab <- @tab do %>
           <li class="mr-2" role="presentation">
             <button
-              class="inline-block p-4 border-b-2 border-transparent rounded-t-lg hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300"
+              class={~w[
+                inline-block p-4 border-b-2 border-transparent rounded-t-lg
+                hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300
+              ]}
               id={"#{tab.id}-tab"}
               data-tabs-target={"##{tab.id}"}
               type="button"
@@ -177,123 +180,36 @@ defmodule Web.CoreComponents do
   def button_group(assigns) do
     ~H"""
     <div class="inline-flex rounded-md shadow-sm" role="group">
-      <button
-        type="button"
-        class="px-4 py-2 text-sm font-medium text-gray-900 bg-white border border-gray-200 rounded-l-lg hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-2 focus:ring-blue-700 focus:text-blue-700 dark:bg-gray-700 dark:border-gray-600 dark:text-white dark:hover:text-white dark:hover:bg-gray-600 dark:focus:ring-blue-500 dark:focus:text-white"
-      >
+      <button type="button" class={~w[
+          px-4 py-2 text-sm font-medium text-gray-900 bg-white border border-gray-200
+          rounded-l-lg hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-2
+          focus:ring-blue-700 focus:text-blue-700 dark:bg-gray-700 dark:border-gray-600
+          dark:text-white dark:hover:text-white dark:hover:bg-gray-600
+          dark:focus:ring-blue-500 dark:focus:text-white
+        ]}>
         <%= render_slot(@first) %>
       </button>
       <%= for middle <- @middle do %>
-        <button
-          type="button"
-          class="px-4 py-2 text-sm font-medium text-gray-900 bg-white border-t border-b border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-2 focus:ring-blue-700 focus:text-blue-700 dark:bg-gray-700 dark:border-gray-600 dark:text-white dark:hover:text-white dark:hover:bg-gray-600 dark:focus:ring-blue-500 dark:focus:text-white"
-        >
+        <button type="button" class={~w[
+            px-4 py-2 text-sm font-medium text-gray-900 bg-white border-t border-b
+            border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-2
+            focus:ring-blue-700 focus:text-blue-700 dark:bg-gray-700 dark:border-gray-600
+            dark:text-white dark:hover:text-white dark:hover:bg-gray-600 dark:focus:ring-blue-500
+            dark:focus:text-white
+          ]}>
           <%= render_slot(middle) %>
         </button>
       <% end %>
-      <button
-        type="button"
-        class="px-4 py-2 text-sm font-medium text-gray-900 bg-white border border-gray-200 rounded-r-lg hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-2 focus:ring-blue-700 focus:text-blue-700 dark:bg-gray-700 dark:border-gray-600 dark:text-white dark:hover:text-white dark:hover:bg-gray-600 dark:focus:ring-blue-500 dark:focus:text-white"
-      >
+      <button type="button" class={~w[
+          px-4 py-2 text-sm font-medium text-gray-900 bg-white border border-gray-200
+          rounded-r-lg hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-2
+          focus:ring-blue-700 focus:text-blue-700 dark:bg-gray-700 dark:border-gray-600
+          dark:text-white dark:hover:text-white dark:hover:bg-gray-600 dark:focus:ring-blue-500
+          dark:focus:text-white
+        ]}>
         <%= render_slot(@last) %>
       </button>
     </div>
-    """
-  end
-
-  @doc """
-  Render a submit button.
-
-  ## Examples
-
-    <.submit_button>
-      Save
-    </.submit_button>
-  """
-
-  slot :inner_block, required: true
-
-  def submit_button(assigns) do
-    ~H"""
-    <button
-      type="submit"
-      class="inline-flex items-center px-5 py-2.5 mt-4 sm:mt-6 text-sm font-medium text-center text-white bg-primary-700 rounded-lg focus:ring-4 focus:ring-primary-200 dark:focus:ring-primary-900 hover:bg-primary-800"
-    >
-      <%= render_slot(@inner_block) %>
-    </button>
-    """
-  end
-
-  @doc """
-  Render a delete button.
-
-  ## Examples
-
-    <.delete_button path={Routes.user_path(@conn, :edit, @user.id)}/>
-      Edit user
-    </.delete_button>
-  """
-  attr :phx_click, :string, doc: "Action to perform when the button is clicked"
-  slot :inner_block, required: true
-
-  def delete_button(assigns) do
-    ~H"""
-    <button
-      type="button"
-      class="text-red-600 inline-flex items-center hover:text-white border border-red-600 hover:bg-red-600 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900"
-    >
-      <!-- XXX: Fix icon for dark mode -->
-      <!-- <.icon name="hero-trash-solid" class="text-red-600 w-5 h-5 mr-1 -ml-1" /> -->
-      <%= render_slot(@inner_block) %>
-    </button>
-    """
-  end
-
-  @doc """
-  Renders an add button.
-
-  ## Examples
-
-    <.add_button navigate={~p"/users/new"}>
-      Add user
-    </.add_button>
-  """
-  attr :navigate, :any, required: true, doc: "Path to navigate to"
-  slot :inner_block, required: true
-
-  def add_button(assigns) do
-    ~H"""
-    <.link
-      navigate={@navigate}
-      class="flex items-center justify-center text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-4 py-2 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800"
-    >
-      <.icon name="hero-plus" class="h-3.5 w-3.5 mr-2" />
-      <%= render_slot(@inner_block) %>
-    </.link>
-    """
-  end
-
-  @doc """
-  Renders an edit button.
-
-  ## Examples
-
-    <.edit_button path={Routes.user_path(@conn, :edit, @user.id)}/>
-      Edit user
-    </.edit_button>
-  """
-  attr :navigate, :any, required: true, doc: "Path to navigate to"
-  slot :inner_block, required: true
-
-  def edit_button(assigns) do
-    ~H"""
-    <.link
-      navigate={@navigate}
-      class="flex items-center justify-center text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-4 py-2 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800"
-    >
-      <.icon name="hero-pencil-solid" class="h-3.5 w-3.5 mr-2" />
-      <%= render_slot(@inner_block) %>
-    </.link>
     """
   end
 
@@ -325,60 +241,66 @@ defmodule Web.CoreComponents do
       </span>
       <ul class="inline-flex items-stretch -space-x-px">
         <li>
-          <a
-            href="#"
-            class="flex items-center justify-center h-full py-1.5 px-3 ml-0 text-gray-500 bg-white rounded-l-lg border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-          >
+          <a href="#" class={~w[
+              flex items-center justify-center h-full py-1.5 px-3 ml-0 text-gray-500 bg-white rounded-l-lg
+              border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700
+              dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white
+            ]}>
             <span class="sr-only">Previous</span>
             <.icon name="hero-chevron-left" class="w-5 h-5" />
           </a>
         </li>
         <li>
-          <a
-            href="#"
-            class="flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-          >
+          <a href="#" class={~w[
+              flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white border
+              border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700
+              dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white
+            ]}>
             1
           </a>
         </li>
         <li>
-          <a
-            href="#"
-            class="flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-          >
+          <a href="#" class={~w[
+              flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white
+              border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700
+              dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white
+            ]}>
             2
           </a>
         </li>
         <li>
-          <a
-            href="#"
-            aria-current="page"
-            class="flex items-center justify-center text-sm z-10 py-2 px-3 leading-tight text-primary-600 bg-primary-50 border border-primary-300 hover:bg-primary-100 hover:text-primary-700 dark:border-gray-700 dark:bg-gray-700 dark:text-white"
-          >
+          <a href="#" aria-current="page" class={~w[
+              flex items-center justify-center text-sm z-10 py-2 px-3 leading-tight text-primary-600 bg-primary-50
+              border border-primary-300 hover:bg-primary-100 hover:text-primary-700 dark:border-gray-700
+              dark:bg-gray-700 dark:text-white
+            ]}>
             <%= @page %>
           </a>
         </li>
         <li>
-          <a
-            href="#"
-            class="flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-          >
+          <a href="#" class={~w[
+              flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white
+              border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700
+              dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white
+            ]}>
             ...
           </a>
         </li>
         <li>
-          <a
-            href="#"
-            class="flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-          >
+          <a href="#" class={~w[
+              flex items-center justify-center text-sm py-2 px-3 leading-tight text-gray-500 bg-white
+              border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700
+              dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white
+            ]}>
             <%= @total_pages %>
           </a>
         </li>
         <li>
-          <a
-            href="#"
-            class="flex items-center justify-center h-full py-1.5 px-3 leading-tight text-gray-500 bg-white rounded-r-lg border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-          >
+          <a href="#" class={~w[
+              flex items-center justify-center h-full py-1.5 px-3 leading-tight text-gray-500 bg-white rounded-r-lg
+              border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700
+              dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white
+            ]}>
             <span class="sr-only">Next</span>
             <.icon name="hero-chevron-right" class="w-5 h-5" />
           </a>
@@ -565,227 +487,6 @@ defmodule Web.CoreComponents do
   end
 
   @doc """
-  Renders a simple form.
-
-  ## Examples
-
-      <.simple_form for={@form} phx-change="validate" phx-submit="save">
-        <.input field={@form[:email]} label="Email"/>
-        <.input field={@form[:username]} label="Username" />
-        <:actions>
-          <.button>Save</.button>
-        </:actions>
-      </.simple_form>
-  """
-  attr :for, :any, required: true, doc: "the datastructure for the form"
-  attr :as, :any, default: nil, doc: "the server side parameter to collect all input under"
-
-  attr :rest, :global,
-    include: ~w(autocomplete name rel action enctype method novalidate target),
-    doc: "the arbitrary HTML attributes to apply to the form tag"
-
-  slot :inner_block, required: true
-  slot :actions, doc: "the slot for form actions, such as a submit button"
-
-  def simple_form(assigns) do
-    ~H"""
-    <.form :let={f} for={@for} as={@as} {@rest}>
-      <div class="space-y-8 bg-white">
-        <%= render_slot(@inner_block, f) %>
-        <div :for={action <- @actions} class="mt-2 flex items-center justify-between gap-6">
-          <%= render_slot(action, f) %>
-        </div>
-      </div>
-    </.form>
-    """
-  end
-
-  @doc """
-  Renders a button.
-
-  ## Examples
-
-      <.button>Send!</.button>
-      <.button phx-click="go" class="ml-2">Send!</.button>
-  """
-  attr :type, :string, default: nil
-  attr :class, :string, default: nil
-  attr :rest, :global, include: ~w(disabled form name value)
-
-  slot :inner_block, required: true
-
-  def button(assigns) do
-    ~H"""
-    <button
-      type={@type}
-      class={[
-        "phx-submit-loading:opacity-75",
-        "text-white bg-primary-600 font-medium rounded-lg text-sm px-5 py-2.5 text-center",
-        "hover:bg-primary-700",
-        "focus:ring-4 focus:outline-none focus:ring-primary-300",
-        "dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800",
-        "active:text-white/80",
-        @class
-      ]}
-      {@rest}
-    >
-      <%= render_slot(@inner_block) %>
-    </button>
-    """
-  end
-
-  @doc """
-  Renders an input with label and error messages.
-
-  A `%Phoenix.HTML.Form{}` and field name may be passed to the input
-  to build input names and error messages, or all the attributes and
-  errors may be passed explicitly.
-
-  ## Examples
-
-      <.input field={@form[:email]} type="email" />
-      <.input name="my-input" errors={["oh no!"]} />
-  """
-  attr :id, :any, default: nil
-  attr :name, :any
-  attr :label, :string, default: nil
-  attr :value, :any
-
-  attr :type, :string,
-    default: "text",
-    values: ~w(checkbox color date datetime-local email file hidden month number password
-               range radio search select tel text textarea time url week)
-
-  attr :field, Phoenix.HTML.FormField,
-    doc: "a form field struct retrieved from the form, for example: @form[:email]"
-
-  attr :errors, :list, default: []
-  attr :checked, :boolean, doc: "the checked flag for checkbox and radio inputs"
-  attr :prompt, :string, default: nil, doc: "the prompt for select inputs"
-  attr :options, :list, doc: "the options to pass to Phoenix.HTML.Form.options_for_select/2"
-  attr :multiple, :boolean, default: false, doc: "the multiple flag for select inputs"
-
-  attr :rest, :global,
-    include: ~w(autocomplete cols disabled form list max maxlength min minlength
-                pattern placeholder readonly required rows size step)
-
-  slot :inner_block
-
-  def input(%{field: %Phoenix.HTML.FormField{} = field} = assigns) do
-    assigns
-    |> assign(field: nil, id: assigns.id || field.id)
-    |> assign(:errors, Enum.map(field.errors, &translate_error(&1)))
-    |> assign_new(:name, fn -> if assigns.multiple, do: field.name <> "[]", else: field.name end)
-    |> assign_new(:value, fn -> field.value end)
-    |> input()
-  end
-
-  def input(%{type: "radio"} = assigns) do
-    ~H"""
-    <div phx-feedback-for={@name}>
-      <label class="flex items-center gap-2 text-gray-900 dark:text-gray-300">
-        <input
-          type="radio"
-          id={@id}
-          name={@name}
-          value={@value}
-          checked={@checked}
-          class="w-4 h-4 border-gray-300 focus:ring-2 focus:ring-blue-300 dark:focus:ring-blue-600 dark:focus:bg-blue-600 dark:bg-gray-700 dark:border-gray-600"
-          {@rest}
-        />
-        <%= @label %>
-      </label>
-    </div>
-    """
-  end
-
-  def input(%{type: "checkbox", value: value} = assigns) do
-    assigns =
-      assign_new(assigns, :checked, fn -> Phoenix.HTML.Form.normalize_value("checkbox", value) end)
-
-    ~H"""
-    <div phx-feedback-for={@name}>
-      <label class="flex items-center gap-4 text-sm leading-6 text-zinc-600">
-        <input type="hidden" name={@name} value="false" />
-        <input
-          type="checkbox"
-          id={@id}
-          name={@name}
-          value="true"
-          checked={@checked}
-          class="rounded border-zinc-300 text-zinc-900 focus:ring-0"
-          {@rest}
-        />
-        <%= @label %>
-      </label>
-      <.error :for={msg <- @errors}><%= msg %></.error>
-    </div>
-    """
-  end
-
-  def input(%{type: "select"} = assigns) do
-    ~H"""
-    <div phx-feedback-for={@name}>
-      <.label for={@id}><%= @label %></.label>
-      <select
-        id={@id}
-        name={@name}
-        class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-        multiple={@multiple}
-        {@rest}
-      >
-        <option :if={@prompt} value=""><%= @prompt %></option>
-        <%= Phoenix.HTML.Form.options_for_select(@options, @value) %>
-      </select>
-      <.error :for={msg <- @errors}><%= msg %></.error>
-    </div>
-    """
-  end
-
-  def input(%{type: "textarea"} = assigns) do
-    ~H"""
-    <div phx-feedback-for={@name}>
-      <.label for={@id}><%= @label %></.label>
-      <textarea
-        id={@id}
-        name={@name}
-        class={[
-          "mt-2 block w-full rounded-lg text-zinc-900 focus:ring-0 sm:text-sm sm:leading-6",
-          "phx-no-feedback:border-zinc-300 phx-no-feedback:focus:border-zinc-400",
-          "min-h-[6rem] border-zinc-300 focus:border-zinc-400",
-          @errors != [] && "border-rose-400 focus:border-rose-400"
-        ]}
-        {@rest}
-      ><%= Phoenix.HTML.Form.normalize_value("textarea", @value) %></textarea>
-      <.error :for={msg <- @errors}><%= msg %></.error>
-    </div>
-    """
-  end
-
-  # All other inputs text, datetime-local, url, password, etc. are handled here...
-  def input(assigns) do
-    ~H"""
-    <div phx-feedback-for={@name}>
-      <.label for={@id}><%= @label %></.label>
-      <input
-        type={@type}
-        name={@name}
-        id={@id}
-        value={Phoenix.HTML.Form.normalize_value(@type, @value)}
-        class={[
-          "mt-2 block w-full rounded-lg text-zinc-900 focus:ring-0 sm:text-sm sm:leading-6",
-          "phx-no-feedback:border-zinc-300 phx-no-feedback:focus:border-zinc-400",
-          "border-zinc-300 focus:border-zinc-400",
-          @errors != [] && "border-rose-400 focus:border-rose-400"
-        ]}
-        {@rest}
-      />
-      <.error :for={msg <- @errors}><%= msg %></.error>
-    </div>
-    """
-  end
-
-  @doc """
   Renders a standard form label.
   """
   attr :for, :string, default: nil
@@ -835,98 +536,6 @@ defmodule Web.CoreComponents do
       </div>
       <div class="flex-none"><%= render_slot(@actions) %></div>
     </header>
-    """
-  end
-
-  @doc ~S"""
-  Renders a table with generic styling.
-
-  ## Examples
-
-      <.table id="users" rows={@users}>
-        <:col :let={user} label="id"><%= user.id %></:col>
-        <:col :let={user} label="username"><%= user.username %></:col>
-      </.table>
-  """
-  attr :id, :string, required: true
-  attr :rows, :list, required: true
-  attr :row_id, :any, default: nil, doc: "the function for generating the row id"
-  attr :row_click, :any, default: nil, doc: "the function for handling phx-click on each row"
-
-  attr :row_item, :any,
-    default: &Function.identity/1,
-    doc: "the function for mapping each row before calling the :col and :action slots"
-
-  slot :col, required: true do
-    attr :label, :string
-    attr :sortable, :string
-  end
-
-  slot :action, doc: "the slot for showing user actions in the last table column"
-
-  def table(assigns) do
-    assigns =
-      with %{rows: %Phoenix.LiveView.LiveStream{}} <- assigns do
-        assign(assigns, row_id: assigns.row_id || fn {id, _item} -> id end)
-      end
-
-    ~H"""
-    <div class="overflow-x-auto">
-      <table class="w-full text-sm text-left text-gray-500 dark:text-gray-400">
-        <thead class="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-          <tr>
-            <th :for={col <- @col} class="px-4 py-3">
-              <%= col[:label] %>
-              <.icon
-                :if={col[:sortable] == "true"}
-                name="hero-chevron-up-down-solid"
-                class="w-4 h-4 ml-1"
-              />
-            </th>
-            <th class="px-4 py-3">
-              <span class="sr-only"><%= gettext("Actions") %></span>
-            </th>
-          </tr>
-        </thead>
-        <tbody id={@id} phx-update={match?(%Phoenix.LiveView.LiveStream{}, @rows) && "stream"}>
-          <tr :for={row <- @rows} id={@row_id && @row_id.(row)} class="border-b dark:border-gray-700">
-            <td
-              :for={{col, _i} <- Enum.with_index(@col)}
-              phx-click={@row_click && @row_click.(row)}
-              class={[
-                "px-4 py-3",
-                @row_click && "hover:cursor-pointer"
-              ]}
-            >
-              <%= render_slot(col, @row_item.(row)) %>
-            </td>
-            <td :if={@action != []} class="px-4 py-3 flex items-center justify-end">
-              <button
-                id={"#{@row_id.(row)}-dropdown-button"}
-                data-dropdown-toggle={"#{@row_id.(row)}-dropdown"}
-                class="inline-flex items-center p-0.5 text-sm font-medium text-center text-gray-500 hover:text-gray-800 rounded-lg focus:outline-none dark:text-gray-400 dark:hover:text-gray-100"
-                type="button"
-              >
-                <.icon name="hero-ellipsis-horizontal" class="w-5 h-5" />
-              </button>
-              <div
-                id={"#{@row_id.(row)}-dropdown" }
-                class="hidden z-10 w-44 bg-white rounded divide-y divide-gray-100 shadow border border-gray-300 dark:bg-gray-700 dark:divide-gray-600"
-              >
-                <ul
-                  class="py-1 text-sm text-gray-700 dark:text-gray-200"
-                  aria-labelledby={"#{@row_id.(row)}-dropdown-button"}
-                >
-                  <li :for={action <- @action}>
-                    <%= render_slot(action, @row_item.(row)) %>
-                  </li>
-                </ul>
-              </div>
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
     """
   end
 
@@ -1088,7 +697,7 @@ defmodule Web.CoreComponents do
   def badge(assigns) do
     colors = %{
       "success" => "bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-300",
-      "failure" => "bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-300",
+      "danger" => "bg-red-100 text-red-800 dark:bg-red-900 dark:text-red-300",
       "warning" => "bg-yellow-100 text-yellow-800 dark:bg-yellow-900 dark:text-yellow-300",
       "info" => "bg-blue-100 text-blue-800 dark:bg-blue-900 dark:text-blue-300",
       "default" => "bg-gray-100 text-gray-800 dark:bg-gray-900 dark:text-gray-300"
@@ -1176,5 +785,37 @@ defmodule Web.CoreComponents do
   """
   def translate_errors(errors, field) when is_list(errors) do
     for {^field, {msg, opts}} <- errors, do: translate_error({msg, opts})
+  end
+
+  @doc """
+  Returns a string the represents a relative time for a given Datetime
+  from the current time or a given base time
+  """
+  attr :relative, DateTime, required: true
+  attr :relative_to, DateTime, required: false, default: DateTime.utc_now()
+
+  def relative_datetime(assigns) do
+    diff = DateTime.diff(assigns[:relative_to], assigns[:relative])
+
+    diff_str =
+      cond do
+        diff <= -24 * 3600 -> "in #{div(-diff, 24 * 3600)}day(s)"
+        diff <= -3600 -> "in #{div(-diff, 3600)}hour(s)"
+        diff <= -60 -> "in #{div(-diff, 60)}minute(s)"
+        diff <= -5 -> "in #{-diff}seconds"
+        diff <= 5 -> "now"
+        diff <= 60 -> "#{diff}seconds ago"
+        diff <= 3600 -> "#{div(diff, 60)} minute(s) ago"
+        diff <= 24 * 3600 -> "#{div(diff, 3600)} hour(s) ago"
+        true -> "#{div(diff, 24 * 3600)} day(s) ago"
+      end
+
+    assigns = assign(assigns, diff_str: diff_str)
+
+    ~H"""
+    <span>
+      <%= @diff_str %>
+    </span>
+    """
   end
 end
